Move token logic to investorView
