+++ containers/Investment/Redeem/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Box, Button, Text } from 'grommet';
 import { redeem, TrancheType } from '../../../services/tinlake/actions';
 import { transactionSubmitted, responseReceived } from '../../../ducks/transactions';
-import { baseToDisplay, displayToBase, Investor, Tranche } from 'tinlake';
+import { baseToDisplay, displayToBase, Investor } from 'tinlake';
 import { loadInvestor } from '../../../ducks/investments';
 import { loadPool } from '../../../ducks/pool';
 import { connect } from 'react-redux';
@@ -16,7 +16,7 @@ interface Props {
   loadPool?: (tinlake: any) => Promise<void>;
   transactionSubmitted?: (loadingMessage: string) => Promise<void>;
   responseReceived?: (successMessage: string | null, errorMessage: string | null) => Promise<void>;
-  tranche: Tranche;
+  tranche: any;
 }
 
 interface State {
@@ -52,39 +52,18 @@ class InvestorRedeem extends React.Component<Props, State> {
     const { investor, tranche } = this.props;
     const { redeemAmount } = this.state;
     const trancheValues = investor[tranche.type as any as TrancheType];
-    const maxRedeemAmount = trancheValues.maxRedeem || '0';
-    const tokenBalance = trancheValues.tokenBalance || '0';
+    const maxRedeemAmount = trancheValues.maxRedeem || new BN('0');
+    const tokenBalance = trancheValues.tokenBalance || new BN('0');
     const redeemLimitSet = maxRedeemAmount.toString() !== '0';
     const limitOverflow = (new BN(redeemAmount).cmp(new BN(maxRedeemAmount)) > 0);
     const availableTokensOverflow = (new BN(redeemAmount).cmp(new BN(tokenBalance)) > 0);
     const redeemEnabled = redeemLimitSet && !limitOverflow && !availableTokensOverflow;
-
-    const dropAddress = this.props.tinlake.contractAddresses.SENIOR_TOKEN as string;
-    const tinAddress = this.props.tinlake.contractAddresses.JUNIOR_TOKEN as string;
-    const dropToken = {
-      [dropAddress] : {
-        symbol: 'DROP',
-        logo: '../../static/DROP_final.svg',
-        decimals: 18,
-        name: 'DROP'
-      }
-    };
-    const tinToken = {
-      [tinAddress] : {
-        symbol: 'TIN',
-        logo: '../../static/TIN_final.svg',
-        decimals: 18,
-        name: 'TIN'
-      }
-    };
+    
     return <Box basis={'1/4'} gap="medium" margin={{ right: 'large' }}>
       <Box gap="medium">
-        {tranche.type === 'senior' && <Erc20Widget input={true} fieldLabel="Redeem token" limit={BN.min(maxRedeemAmount,tokenBalance).toString()} tokenData={dropToken} precision={18} onValueChanged={(value : string) =>
+        <Erc20Widget input={true} fieldLabel="Redeem token" limit={BN.min(maxRedeemAmount,tokenBalance).toString()} tokenData={tranche.tokenData} precision={18} onValueChanged={(value : string) =>
                 this.setState({ redeemAmount: displayToBase(value, 18) })}
-                errorMessage="Max redeem amount exceeded" />}
-        {tranche.type === 'junior' && <Erc20Widget input={true} fieldLabel="Redeem token" limit={BN.min(maxRedeemAmount,tokenBalance).toString()} tokenData={tinToken} precision={18} onValueChanged={(value : string) =>
-                this.setState({ redeemAmount: displayToBase(value, 18) })}
-                errorMessage="Max redeem amount exceeded" />}
+                errorMessage="Max redeem amount exceeded" />
       </Box>
       <Box align="start">
         <Button onClick={this.redeem} primary label="Redeem" disabled = {!redeemEnabled}/>
